mail/requirements.txt
absl-py==1.4.0
aspose-words==23.2.0
Aspose.Slides==23.1.0
astunparse==1.6.3
attrs==22.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
contextlib2==21.6.0
exceptiongroup==1.1.0
execnet==1.9.0
flatbuffers==23.1.21
gast==0.4.0
google==3.0.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.51.3
h5py==3.8.0
httplib2==0.21.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
keras==2.11.0
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
mock==5.0.1
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
path==16.6.0
path.py==12.5.0
Pillow==9.4.0
pluggy==1.0.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytest==7.2.1
pytest-shutil==1.7.0
python-magic==0.4.27
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tomli==2.0.1
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.14.0
